requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.7.7
openai==1.3.5
pygetwindow==0.0.9
pyautogui==0.9.54
psutil==5.9.6
pillow>=11.0.0
python-dotenv==1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings==2.13.1
httpx==0.25.1
pyperclip==1.8.2
